fix: add missing abc import
