+++ pyproject.toml
@@ -1,8 +1,10 @@
 [project]
 name = "agent-debugger"
 version = "0.1.0"
-description = "Add your description here"
+description = "Visual interface for debugging AI agent workflows"
 readme = "README.md"
+authors = [{name = "Agent Debugger Contributors"}]
+license = {text = "MIT"}
 requires-python = ">=3.12"
 dependencies = [
     "fastapi>=0.116.1",
