+++ run_fastapi.py
@@ -3,6 +3,7 @@
 
 import uvicorn
 
+
 if __name__ == "__main__":
     # 运行FastAPI应用
     uvicorn.run(
@@ -10,5 +11,6 @@ if __name__ == "__main__":
         host="0.0.0.0",
         port=8000,
         reload=True,
-        log_level="info"
+        log_level="info",
     )
+
