+++ tools/base.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
-from typing import Any, Dict, Optional, List
 from dataclasses import dataclass
+from typing import Any, Dict, List, Optional
 
 
 @dataclass
